requirements.txt
aiofiles==0.4.0
aiohttp==3.7.4
aiohttp-socks==0.5.5
async-timeout==3.0.1
attrs==20.3.0
certifi==2020.12.5
chardet==3.0.4
click==7.1.2
fastapi==0.63.0
future==0.18.2
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hyperframe==5.2.0
idna==2.10
jsonschema==3.2.0
Logbook==1.5.3
matrix-nio==0.17.0
multidict==5.1.0
pycryptodome==3.10.1
pydantic==1.8.1
pyrsistent==0.17.3
python-socks==1.2.2
requests==2.25.1
six==1.15.0
starlette==0.13.6
typing-extensions==3.7.4.3
unpaddedbase64==1.1.0
urllib3==1.26.3
uvicorn==0.13.4
yarl==1.6.3
